refactor: Проведен рефакторинг компонента orders и вложенных компонентов
